fix EDA write bug
