+++ rdagent/components/coder/data_science/raw_data_loader/__init__.py
@@ -231,5 +231,9 @@ class DataLoaderCoSTEER(CoSTEER):
         stdout = new_exp.experiment_workspace.execute(env=env, entry=f"python test/data_loader_test.py")
         match = re.search(r"(.*?)=== Start of EDA part ===(.*)=== End of EDA part ===", stdout, re.DOTALL)
         eda_output = match.groups()[1] if match else None
-        new_exp.experiment_workspace.inject_files(**{"EDA.md": eda_output})
+        if eda_output is not None:
+            new_exp.experiment_workspace.inject_files(**{"EDA.md": eda_output})
+        else:
+            eda_output = "No EDA output."
+            new_exp.experiment_workspace.inject_files(**{"EDA.md": eda_output})
         return new_exp
